change focal point labels
if 2 or more focal points were really close (not perfectly on top of each other but close), the labels would overlap. The variable verticalSpace is added to the drawPointsOfInterest. Each time a z has already be append to the listOfZ array (a point has been drawn), and if the new z is within 20% of the total x range of the plot, the label will be drawn a certain amount below the normal position. It is incrementable.

I added a variable called idxLens to label the point of interest with its corresponding lens.